chore: run serialization tests on all `Format` variants
