+++ crates/noirc_abi/Cargo.toml
@@ -13,3 +13,7 @@ toml.workspace = true
 serde_json = "1.0"
 serde.workspace = true
 thiserror.workspace = true
+
+[dev-dependencies]
+strum = "0.24"
+strum_macros = "0.24"
